Fix "Module load not found"
We separated require_priv in functions but by that we broke the sizeof()
resolution.  If you pass a pointer to a function and call sizeof() on that
it will return the size of the pointer type itself instead of the stack
allocated size.  This change fixes it by using the allocated size
instead of calling sizeof().  Although magic numbers should not be used
PATH_MAX makes sense as we are handling PATHS.