E3-1589: Fix truncated error logs
In EPICS base, the errorlog functionality is initialised with a default
maximum error log size of 256. This means that error messages longer than
that size show as `<<TRUNCATED>>´, which means that we miss a number of
important messages, in particular error messages due to missing symbols
from shared libraries when they are loaded.

errlogInit unfortunately is a bit of a singleton; it can only be called
a single time, and it is called when we load require the very first time.