Overwrite where_am_I variable
This variable was historically defined in the module.makefiles via the
boilerplate code
```make
where_am_I := $(dir $(abspath $(lastword $(MAKEFILE_LIST))))
include $(E3_REQUIRE_TOOLS)/driver.makefile
```
This was thus defined as the module directory: it is the absolute path of
the directory that the module makefile sits in, which was copied into the
module directory.

We can simply overwrite this as `CURDIR` (give or take) which is a make
variable that defines the directory we are currently running in.

It might be worth cleaning this up somewhat on a per-module basis, but this
does avoid actually having to do a bulk modification.