Rework on module load and register with init.cpp
This change simplify require by off loading the function of loading
dependent modules to ld. Instead of finding the dependency modules and
loading them one by one require will only need to load the first one and
ld will load all the dependencies.

The main mechanism that permits this is by moving the
<module>_registerRecordDeviceDriver and registerModule calls to the
module itself.

An important "hack" is used on this change to avoid undefined behavior
by the static initialization order fiasco. To fix this the init.cpp is
appended to the end of module_registerRecordDeviceDriver.cpp and
Registration is called directly by __module_library_init().