Port old entries of the CHANGELOG.md from e3-require
Most of the development from previous years were on the e3-require fork.
Five versions were released from there and the code later was imported
here however the CHANGELOG was not updated correctly. This commit ported
the entries from e3-require and try to consolidate so it makes sense for
the releases done from this repository.