Move TMPS/SUBS to outer loop
Note that we have to change variable name (see other examples) so that
they are not doubled up on the second loop.

We have to keep USR_DBFLAGS in the second loop, however, since we want
to avoid doubling _and_ we want to keep the name consistent with EPICS
base.